requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asgi-lifespan==2.1.0
asyncpg==0.30.0
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flake8==7.2.0
Flake8-pyproject==1.2.3
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.0
hyperframe==6.1.0
identify==2.6.9
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
orjson==3.10.16
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
pycodestyle==2.13.0
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyflakes==3.3.2
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.30.0
watchfiles==1.0.5
websockets==15.0.1
